move (r)info cmd tests to own module
